move tests into types_tests
